+++ selfserve-jobs-customer-web/next.config.ts
@@ -17,11 +17,7 @@ const nextConfig: NextConfig = {
     remotePatterns: [
       {
         protocol: 'https',
-        hostname: '**',
-      },
-      {
-        protocol: 'http',
-        hostname: '**',
+        hostname: 'storage.googleapis.com',
       },
     ],
   },